Initial commit
Essentially a proof of concept. Allows you to configure this module by
pointing it at templates. By running ":Neorge capture" you can select
which template to use, and then a new window will split and populate the
buffer with that template. When you save and quit that buffer (not just
save) it will write it to the end of the `file` configuation variable.

This is very rough.