Target by Query, refactoring, and bug fixes.
If the target file is already open in a buffer, it will now use that
buffer instead of opening a new one and reading into it. This is to
prevent errors of the type of "file already open" when trying to write.

Did some refactoring just to reduce and reuse code. During that
processes I realized how simple it would be to implement the abililty to
target a headline (or anywhere else in an norg document I presume) for a
user by defining a treesitter query. This is a feature I had thought
about and wanted to eventually add, but after the refactor it became a
very simple thing to implement so I went ahead.